# static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"RegStudentDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"referrer_id": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name",
					"phone_number",
					"email",
					"password",
					"referrer_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TStudent": {
				"allOf": [
					{
						"$ref": "#/components/schemas/RegStudentDTO"
					},
					{
						"properties": {
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					}
				]
			},
			"AuthStudentDTO": {
				"properties": {
					"phone_or_email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"phone_or_email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TStudentLesson": {
				"properties": {
					"transaction_id": {
						"type": "string"
					},
					"lesson_id": {
						"type": "string"
					},
					"student_id": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"transaction_id",
					"lesson_id",
					"student_id",
					"id"
				],
				"type": "object"
			},
			"PayLessonDTO": {
				"properties": {
					"lesson_id": {
						"type": "string"
					}
				},
				"required": [
					"lesson_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TStatistic": {
				"properties": {
					"amount": {
						"type": "string"
					},
					"referral_id": {
						"type": "string"
					}
				},
				"required": [
					"amount",
					"referral_id"
				],
				"type": "object"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "referral",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/students": {
			"post": {
				"operationId": "RegisterStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TStudent"
								}
							}
						}
					}
				},
				"tags": [
					"Students"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegStudentDTO"
							}
						}
					}
				}
			}
		},
		"/students/auth": {
			"post": {
				"operationId": "AuthorizeStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"name": {
											"type": "string"
										},
										"_id": {
											"type": "string"
										},
										"token": {}
									},
									"required": [
										"name",
										"_id",
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Students"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthStudentDTO"
							}
						}
					}
				}
			}
		},
		"/students/referral": {
			"get": {
				"operationId": "GenReferralLink",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"link": {
											"type": "string"
										}
									},
									"required": [
										"link"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Students"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/transactions/payment": {
			"post": {
				"operationId": "PayForLesson",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TStudentLesson"
								}
							}
						}
					}
				},
				"tags": [
					"Transactions"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PayLessonDTO"
							}
						}
					}
				}
			}
		},
		"/transactions/statistic": {
			"get": {
				"operationId": "GetReferralStatistic",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TStatistic"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Transactions"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}